backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
openstacksdk
flask-cors
Werkzeug